refactor

+++ services/ad.js
@@ -1,19 +1,38 @@
 const moment = require('moment')
+const mongoose = require('mongoose')
+
 const Ads = require('../model/ad')
 const { querySortValidator, queryAdValidator, updateAdlidator } = require('../utils')
 const { PAGE_SIZE, DICTIONARY, PARSED_OBJECTS, DATE_FORMAT } = require('../constants')
 
 exports.getAd = async (adId, query) => {
   try {
-    if (Object.keys(query).length === 0) {
-      const ad = await Ads.findById(adId, PARSED_OBJECTS.withoutParams)
-      return ad
-    } else if (query.fields && Object.keys(query).length === 1) {
-      const isValid = queryAdValidator(query.fields)
-      if (isValid) {
-        return await Ads.findById(adId, isValid)
+    if (Object.keys(query).length === 0 && adId) {
+      if (mongoose.Types.ObjectId.isValid(adId)) {
+        const ad = await Ads.findById(adId, PARSED_OBJECTS.withoutParams)
+        if (ad === null) {
+          throw new Error(DICTIONARY.errors.adNotFound)
+        } else {
+          return ad
+        }
       } else {
-        throw new Error(DICTIONARY.errors.badRequest)
+        throw new Error(DICTIONARY.errors.badId)
+      }
+    } else if (query.fields && adId && Object.keys(query).length === 1) {
+      if (mongoose.Types.ObjectId.isValid(adId)) {
+        const isValid = queryAdValidator(query.fields)
+        if (isValid) {
+          const ad = await Ads.findById(adId, isValid)
+          if (ad === null) {
+            throw new Error(DICTIONARY.errors.adNotFound)
+          } else {
+            return ad
+          }
+        } else {
+          throw new Error(DICTIONARY.errors.badFields)
+        }
+      } else {
+        throw new Error(DICTIONARY.errors.badID)
       }
     } else {
       throw new Error(DICTIONARY.errors.badRequest)
@@ -25,25 +44,29 @@ exports.getAd = async (adId, query) => {
 
 exports.getAds = async query => {
   try {
-    const isValid = querySortValidator(query)
-    if (isValid && isValid.hasOwnProperty('date')) {
-      const ads = await Ads
-        .find({}, PARSED_OBJECTS.withDate)
-        .skip(PAGE_SIZE * (query.page - 1))
-        .limit(PAGE_SIZE)
-        .sort(isValid)
-      return ads.map(a => ({
-        title: a.title,
-        price: a.price,
-        mainUrl: a.mainUrl,
-        date: moment(a.date).format(DATE_FORMAT)
-      }))
-    } else if (isValid) {
-      return await Ads
-        .find({}, PARSED_OBJECTS.withoutParams)
-        .skip(PAGE_SIZE * (query.page - 1))
-        .limit(PAGE_SIZE)
-        .sort(isValid)
+    if (Object.keys(query).length === 2 && query.sort && query.page) {
+      const isValid = querySortValidator(query)
+      if (isValid && isValid.hasOwnProperty('date')) {
+        const ads = await Ads
+          .find({}, PARSED_OBJECTS.withDate)
+          .skip(PAGE_SIZE * (query.page - 1))
+          .limit(PAGE_SIZE)
+          .sort(isValid)
+        return ads.map(a => ({
+          title: a.title,
+          price: a.price,
+          mainUrl: a.mainUrl,
+          date: moment(a.date).format(DATE_FORMAT)
+        }))
+      } else if (isValid) {
+        return await Ads
+          .find({}, PARSED_OBJECTS.withoutParams)
+          .skip(PAGE_SIZE * (query.page - 1))
+          .limit(PAGE_SIZE)
+          .sort(isValid)
+      } else {
+        throw new Error(DICTIONARY.errors.badSortFields)
+      }
     } else {
       throw new Error(DICTIONARY.errors.badRequest)
     }
@@ -63,11 +86,20 @@ exports.createAd = async ad => {
 
 exports.updateAd = async (adId, body) => {
   try {
-    const isValid = updateAdlidator(body)
-    if (isValid) {
-      return await Ads.findOneAndUpdate({ _id: adId }, body, { new: true })
+    if (mongoose.Types.ObjectId.isValid(adId)) {
+      const isValid = updateAdlidator(body)
+      if (isValid) {
+        const newAd = await Ads.findOneAndUpdate({ _id: adId }, body, { new: true })
+        if (newAd === null) {
+          throw new Error(DICTIONARY.errors.adNotFound)
+        } else {
+          return newAd
+        }
+      } else {
+        throw new Error(DICTIONARY.errors.badBody)
+      }
     } else {
-      throw new Error(DICTIONARY.errors.badRequest)
+      throw new Error(DICTIONARY.errors.badID)
     }
   } catch (err) {
     return err
@@ -76,11 +108,15 @@ exports.updateAd = async (adId, body) => {
 
 exports.deleteAd = async adId => {
   try {
-    const deleteAd = await Ads.findOneAndDelete({ _id: adId })
-    if (deleteAd === null) {
-      throw new Error('Ad not exist')
+    if (mongoose.Types.ObjectId.isValid(adId)) {
+      const deleteAd = await Ads.findOneAndDelete({ _id: adId })
+      if (deleteAd === null) {
+        throw new Error(DICTIONARY.errors.adNotFound)
+      }
+      return deleteAd
+    } else {
+      throw new Error(DICTIONARY.errors.badID)
     }
-    return deleteAd
   } catch (err) {
     return err
   }
